Total refactor of tm processing
* Created a product factory to hanel opening readig all products
* Created methods on the highlevel data to create from low leve data e.g l1.from_level0(l0)
