+++ stixcore/io/fits/processors.py
@@ -6,8 +6,11 @@ import numpy as np
 
 import astropy.units as u
 from astropy.io import fits
+from astropy.io.fits import table_to_hdu
 from astropy.table import QTable
 
+from stixcore.products.common import _get_energies_from_mask
+from stixcore.products.product import Product
 from stixcore.util.logging import get_logger
 
 __all__ = ['SEC_IN_DAY', 'FitsProcessor', 'FitsLBProcessor']
@@ -43,7 +46,8 @@ class FitsProcessor:
             user_req = f'_{user_req}'
 
         scet_obs = (product.control["scet_coarse"][0] // SEC_IN_DAY) * SEC_IN_DAY
-        name = "-".join([str(x) for x in (getattr(product, 'type', ()))])
+        name = '-'.join([str(x) for x in [product.service_type, product.service_subtype,
+                                          product.ssid]])
         return f'solo_{product.level}_stix-{name}' \
                f'_{scet_obs}_V{version:02d}{status}.fits'
 
@@ -73,14 +77,15 @@ class FitsProcessor:
             ('OBT_BEG', f'{product.control["scet_coarse"][0]}:{product.control["scet_fine"][0]}'),
             ('OBT_END', f'{product.control["scet_coarse"][-1]}:{product.control["scet_fine"][-1]}'),
             ('TIMESYS', 'OBT', 'System used for time keywords'),
-            ('LEVEL', 'LB)', 'Processing level of the data'),
+            ('LEVEL', 'LB', 'Processing level of the data'),
             ('ORIGIN', 'STIX Team, FHNW', 'Location where file has been generated'),
             ('CREATOR', 'STIX-SWF', 'FITS creation software'),
             ('VERSION', 1, 'Version of data product'),
             ('OBS_MODE', 'Nominal'),
             ('VERS_SW', 1, 'Software version'),
-            ('TYPE', "-".join([str(x) for x in (getattr(product, 'type', ()))]),
-             'Packet Type service-subservice[-ssid])')
+            ('STYPE', product.service_type),
+            ('SSTYPE', product.service_subtype),
+            ('SSID', product.ssid)
         )
         return headers
 
@@ -112,15 +117,14 @@ class FitsLBProcessor(FitsProcessor):
         """
         for prod in product.to_days():
             filename = self.generate_filename(prod, version=1)
-            parts = [prod.level]
-            parts.extend(prod.type)
+            parts = [prod.level, prod.service_type, prod.service_subtype, prod.ssid]
             path = self.archive_path.joinpath(*[str(x) for x in parts])
             path.mkdir(parents=True, exist_ok=True)
             fitspath = path / filename
             if fitspath.exists():
                 logger.info('Fits file %s exists appending data', fitspath.name)
-                existing = prod.from_fits(fitspath)
-                existing.type = prod.type
+                existing = Product(fitspath)
+                # existing.type = prod.type
                 if np.abs([((len(existing.data['data'][i])/2) -
                             (existing.control['data_length'][i]+7))
                           for i in range(len(existing.data))]).sum() > 0:
@@ -190,7 +194,7 @@ class FitsL0Processor:
             # start_day = np.floor((prod.obs_beg.as_float()
             #                       // (1 * u.day).to('s')).value * SEC_IN_DAY).astype(int)
             parts = ['L0', prod.service_type, prod.service_subtype,
-                     prod.pi1_val]
+                     prod.ssid]
             path = self.archive_path.joinpath(*[str(x) for x in parts])
             path.mkdir(parents=True, exist_ok=True)
 
@@ -198,7 +202,7 @@ class FitsL0Processor:
             fitspath = path / filename
             if fitspath.exists():
                 logger.info('Fits file %s exists appending data', fitspath.name)
-                existing = prod.from_fits(fitspath)
+                existing = Product(fitspath)
                 logger.debug('Existing %s \n Current %s', existing, prod)
                 prod = prod + existing
                 logger.debug('Combined %s', prod)
@@ -206,12 +210,12 @@ class FitsL0Processor:
             control = prod.control
             data = prod.data
 
-            elow, ehigh = prod.get_energies()
-
-            energies = QTable()
-            energies['channel'] = range(len(elow))
-            energies['e_low'] = elow * u.keV
-            energies['e_high'] = ehigh * u.keV
+            # elow, ehigh = prod.get_energies()
+            #
+            # energies = QTable()
+            # energies['channel'] = range(len(elow))
+            # energies['e_low'] = elow * u.keV
+            # energies['e_high'] = ehigh * u.keV
 
             # Convert time to be relative to start date
             data['time'] = (data['time'] - prod.obs_beg.as_float()).to(u.s)
@@ -221,14 +225,17 @@ class FitsL0Processor:
             primary_hdu.header.update(primary_header)
             primary_hdu.header.update({'HISTORY': 'Processed by STIX'})
 
-            control_hdu = fits.BinTableHDU(control)
+            control_enc = fits.connect._encode_mixins(control)
+            control_hdu = table_to_hdu(control_enc)
             control_hdu.name = 'CONTROL'
-            data_hdu = fits.BinTableHDU(data)
+            data_enc = fits.connect._encode_mixins(data)
+            data_hdu = table_to_hdu(data_enc)
             data_hdu.name = 'DATA'
-            energy_hdu = fits.BinTableHDU(energies)
-            energy_hdu.name = 'ENERGIES'
+            # energy_enc = fits.connect._encode_mixins(energies)
+            # energy_hdu = table_to_hdu(energy_enc)
+            # energy_hdu.name = 'ENERGIES'
 
-            hdul = fits.HDUList([primary_hdu, control_hdu, data_hdu, energy_hdu])
+            hdul = fits.HDUList([primary_hdu, control_hdu, data_hdu])  # , energy_hdu])
 
             logger.debug(f'Writing fits file to {path / filename}')
             hdul.writeto(path / filename, overwrite=True, checksum=True)
@@ -299,7 +306,7 @@ class FitsL0Processor:
              'FITS file creation date in UTC'),
             ('OBT_BEG', str(product.obs_beg)),
             ('OBT_END', str(product.obs_end)),
-            ('TIMESYS', 'UTC', 'System used for time keywords'),
+            ('TIMESYS', 'OBT', 'System used for time keywords'),
             ('LEVEL', 'L0', 'Processing level of the data'),
             ('ORIGIN', 'STIX Team, FHNW', 'Location where file has been generated'),
             ('CREATOR', 'STIX-SWF', 'FITS creation software'),
@@ -320,8 +327,7 @@ class FitsL0Processor:
             ('TARGET', 'Sun'),
             ('STYPE', product.service_type),
             ('SSTYPE', product.service_subtype),
-            ('SSID', product.pi1_val),
-            ('SPID', product.spid)
+            ('SSID', product.ssid)
         )
         return headers
 
@@ -358,7 +364,12 @@ class FitsL1Processor:
 
             control = prod.control
             data = prod.data
-            energies = prod.energies
+
+            elow, ehigh = _get_energies_from_mask(control['energy_bin_edge_mask'][0])
+            energies = QTable()
+            energies['channel'] = range(len(elow))
+            energies['e_low'] = elow * u.keV
+            energies['e_high'] = ehigh * u.keV
 
             # Convert time to be relative to start date
             data['time'] = (data['time'] - prod.obs_beg).to(u.s)
@@ -476,7 +487,6 @@ class FitsL1Processor:
             ('TARGET', 'Sun'),
             ('STYPE', product.service_type),
             ('SSTYPE', product.service_subtype),
-            ('SSID', product.pi1_val),
-            ('SPID', product.spid)
+            ('SSID', product.ssid)
         )
         return headers
