+++ stixcore/products/__init__.py
@@ -1,0 +1,4 @@
+from stixcore.products.level0.quicklook import LightCurve
+from stixcore.products.level1.quicklook import LightCurve
+from stixcore.products.levelb.binary import LevelB
+from stixcore.products.product import Product
