+++ stixcore/products/tests/test_quicklook.py
@@ -1,13 +1,13 @@
 from binascii import unhexlify
 
-from stixcore.products.quicklook import LightCurve
+from stixcore.products.level0.quicklook import LightCurve
 from stixcore.tmtc import Packet
 from stixcore.tmtc.packets import PacketSequence
 
 
 def test_lightcurve():
     ql_lc_packets = PacketSequence([Packet(unhexlify('0ddc83fc068510150600274db0d4cd8f1e274dad2e00080027fffffcff0fff2b56108210820500e83231323231313131313131313030313231323132313032303131303031313131313231303131313231323232313231323231323232303131323131313131313130313131323132313131313132313131313132313131313132313131333031313131313131313131313132303131313131313132313132313132323131313131313131303131313233313131303132323231313131313031313130313131313131313131313132323130313131313132323030313230313231313131313130313130303231313131313031303131313031313031313031323131303131313131323030323130313100e81f1c21211f1f1f201f2122221e2121211c202023191f211e21201d202120211c20212021201f20201e202020222020201e20211f1e201f201f211c1f1b202020201b201e1a20231f2220201d2020211f1f2020212220211d2021201e202020201e1f1c1d1f1f1d231f1f222121211f212022202221201f201f1f22221f1d21202020202121201f2120201e20202222211f21202021201e1f20212220202022211f211f21212020202020201e20202121202120211e212023212021211f1e1f1d21201f201e20202020231f1f1d212121201f2021201e2020201e2020211d20231f202123201e211f00e81f202120212221201e1f1d1f1f2020201e2420211f2320201c201f2220201e211f211f20211f2021202120201e211f21201e221e231f1f1d1f1d1f22201f1f1f1f211f20202121201b1f1e1f1e2020201e201e1f1e1e1e1f20232120211f1d201f22221d21211e201f1f211d211f2021202120221f1f202120201f1f211d1e1f1f1e21202022211f1f1f221e1d20202121201f211e1e1e1c1d1e1e1d1e1f201e211e1c201d1f1d212020201d1d20201f1f202021201e1d1f211f1f1e1c211f22201e20211e202020202022201f2120211f1d1f1f1f2020201f1e20201e1d1f2220211f1e211f1d2100e83d3c3d3d3e3e3e3e3d3d3e3d3d3c3d3d3c3d3d3e3d3d3d3d3d3d3d3d3e3e3d3e3d3e3e3d3c3d3e3c3d3d3e3d3d3d3c3c3d3d3d3d3d3d3c3d3d3d3e3e3c3e3e3d3d3e3d3d3d3d3c3d3d3d3c3d3d3d3e3d3d3e3d3d3e3c3d3c3e3d3d3d3d3e3c3d3d3d3d3d3d3e3e3d3d3d3d3e3e3d3d3e3e3d3e3d3d3d3e3e3d3d3d3d3e3d3d3d3d3d3d3e3d3d3d3d3e3e3e3d3d3d3e3c3d3d3d3e3e3d3d3d3d3e3d3d3d3d3d3c3d3d3d3d3d3d3d3d3e3d3e3d3d3e3d3c3d3d3d3d3c3d3d3d3d3d3d3e3e3e3d3e3d3d3d3d3d3d3d3e3d3d3d3d3e3e3d3e3d3e3d3d3d3d3d3e3e3d3d3e3d3c3e3e3e3d3d3d3d3d3d3d00e83736373837363636373536373635373736373736373637353636363736373637363636373836363637363637373737363637363637373637363636363736363637363637373536363636373636363737363737363737373736353737373736363637363735363737363837373736363637373737363537363637373536363736373737363736373637363836363636363736373537373737373636373536383836363637373735363737363737363736373735373738383634363737353737363736363638363737363636363636373536373636373535383738373637373637373637383736363600e8504f5151505150525151504f4f50515050505051515050514f51505251505051505050515051505050505050515151505150505151514f5050515151505051515051505050505150515051505050515051505150514f51505150515052505051505250505050515051515051505250515150515051505150515250515150505050525050525151515152514f5052505150505050505151505050515151515051515150504f5050515151505150505151505150515051505050525051515150515150515150505151505150505151515150505050525050515151505051505152515150505250514f00e800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000'))])  # NOQA
-    ql_lc_product = LightCurve.from_packets(ql_lc_packets)
+    ql_lc_product = LightCurve(ql_lc_packets)
     assert ql_lc_product.level == 'L0'
     assert ql_lc_product.name == 'LightCurve'
     assert len(ql_lc_product.control) == 1
